+++ influxdb_client/client/users_api.py
@@ -6,7 +6,7 @@ and provide them with an authentication token.
 """
 
 from typing import Union
-from influxdb_client import UsersService, User, Users, UserResponse
+from influxdb_client import UsersService, User, Users, UserResponse, PasswordResetBody
 
 
 class UsersApi(object):
@@ -36,26 +36,32 @@ class UsersApi(object):
         """
         return self._service.patch_users_id(user_id=user.id, user=user)
 
+    def update_password(self, user: Union[str, User, UserResponse], password: str) -> None:
+        """Update a password.
+
+        :param user: User to update password (required)
+        :param password: New password (required)
+        :return: None
+        """
+        user_id = self._user_id(user)
+
+        return self._service.post_users_id_password(user_id=user_id, password_reset_body=PasswordResetBody(password))
+
     def delete_user(self, user: Union[str, User, UserResponse]) -> None:
         """Delete a user.
 
         :param user: user id or User
-        :return: User
+        :return: None
         """
-        if isinstance(user, User):
-            user_id = user.id
-        elif isinstance(user, UserResponse):
-            user_id = user.id
-        else:
-            user_id = user
+        user_id = self._user_id(user)
 
         return self._service.delete_users_id(user_id=user_id)
 
     def find_users(self, **kwargs) -> Users:
         """List all users.
 
-        :key int offset: Offset for pagination
-        :key int limit: Limit for pagination
+        :key int offset: The offset for pagination. The number of records to skip.
+        :key int limit: Limits the number of records returned. Default is `20`.
         :key str after: The last resource ID from which to seek from (but not including).
                         This is to be used instead of `offset`.
         :key str name: The user name.
@@ -63,3 +69,12 @@ class UsersApi(object):
         :return: Buckets
         """
         return self._service.get_users(**kwargs)
+
+    def _user_id(self, user: Union[str, User, UserResponse]):
+        if isinstance(user, User):
+            user_id = user.id
+        elif isinstance(user, UserResponse):
+            user_id = user.id
+        else:
+            user_id = user
+        return user_id